requirements.txt
aiohttp>=3.6.0
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
discord.py==2.3.2
dnspython==1.15.0
dnspython3==1.15.0
idna==2.9
mcstatus==3.1.0
multidict==4.7.5
py-dactyl==0.1.5
pycodestyle==2.5.0
requests==2.23.0
six==1.14.0
urllib3!=1.25.0
websockets==8.1
yarl==1.4.2
